==> 03 Scripts/1. Instacart Data Cleaning.ipynb <==
# Import libraries & dataframes

# import libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
from matplotlib.ticker import MaxNLocator
import plotly.express as px
from scipy.stats import ttest_ind
from scipy.stats import ttest_1samp
import seaborn as sns

# import dataframes
path = r'C:\Users\muril\Data Analysis Projects\07-2023 Instacart Basket Analysis\02 Data'

ords = pd.read_csv(os.path.join(path,'Original Data','orders.csv'))

prods = pd.read_csv(os.path.join(path,'Original Data','products.csv'))

dept = pd.read_csv(os.path.join(path,'Original Data','departments.csv'))

cust = pd.read_csv(os.path.join(path,'Original Data','customers.csv'))

# 1. Orders

# check dimensions of orders
ords.shape

## 1.1 Wrangling

# view first rows of dataframe
ords.head()

### 1.1.1 Datatypes

# check for (mixed) datatypes
ords.dtypes

order_id and user_id can be converted to strings, since they represent names rather than quantifiable data.

# convert order_id and user_id to strings
ords['order_id'] = ords['order_id'].astype(str)
ords['user_id'] = ords['user_id'].astype(str)

# check for (mixed) datatypes
ords.dtypes

### 1.1.2 Columns to drop

# investigate eval_set column
ords['eval_set'].value_counts()

On verification with the client, this column is not useful to the project and can be ignored, hence it will be dropped.

# drop eval_set
ords = ords.drop('eval_set', axis=1)

### 1.1.3 Columns to rename

# order_dow isn't intuitive and, on verification with the client, means order_day_of_week and so will be renamed
ords.rename(columns={'order_dow': 'order_day_of_week'}, inplace=True)

# review dataframe
ords.head()

ords.dtypes

Wrangling steps on orders set complete.

## 1.2 Consistency checks

### 1.2.1 Duplicates

# check for duplicate values
ords.duplicated().sum()

There are no duplicate rows.

### 1.2.2 Missing values

# check for nulls
ords.isnull().sum()

# investigate days_since_prior_order column
ords[ords['days_since_prior_order'].isnull()]

# count the number of users
ords['user_id'].nunique()

# find how many users have more than 1 null in the days_since_prior_order column
( ords.groupby('user_id')['days_since_prior_order'].apply(lambda x: x.isnull().sum()) >1 ).sum()

There is one null for every user in the days_since_prior_order column. Verification with the client confirms that this is because it is the first order made by the user. These nulls are not true missing values and will be left unchanged.

There are no missing values in the orders dataset.

### 1.2.3. Inaccuracies

# find summary statistics for the dataset
ords.describe()

The counts are all correct for the dataset.

None of the summary statistics are suspicious, therefore we will proceed on the basis that are no errors in the dataset.

# 2. Products

# view dimensions of products dataset
prods.shape

# view dataset
prods.head()

## 2.1 Wrangling

### 2.1.1 Datatypes

# check for mixed datatypes
prods.dtypes

product_id, aisle_id and department_id can be converted to strings, since they represent names rather than quantifiable data.

# convert product_id, aisle_id and department_id to strings
prods['product_id'] = prods['product_id'].astype(str)
prods['aisle_id'] = prods['aisle_id'].astype(str)
prods['department_id'] = prods['department_id'].astype(str)

# check datatypes
prods.dtypes

There are no mixed datatypes to fix.

### 2.1.2 Columns to drop

There are no columns that need to be dropped.

### 2.1.3 Columns to rename

All column names are clear and succinct, no need to change any names.

## 2.2 Consistency checks

### 2.2.1 Duplicates

# check for duplicate values
prods.duplicated().sum()

# view duplicates
prods[prods.duplicated(keep='last')]

# view counts of these rows
prods[prods['product_id'].isin([462,18458,26808,35306,35491])].value_counts()

Each of these rows are duplicated once. Drop them.

# drop duplicates
prods.drop_duplicates(inplace=True)

# check for duplicate values
prods.duplicated().sum()

# check new dimensions
prods.shape

Duplicates removed.

## 2.2.2 Missing values

# check for nulls
prods.isnull().sum()

# view rows with missing values
prods[prods['product_name'].isnull()]

All numerical data is intact and looks sensible. Since we have the product_id, request the information from the client to be added to the dataset and do not drop the rows, as this will unnecessarily distort the analysis.

## 2.2.3 Inaccuracies

# find summary statistics for the dataset
prods.describe()

Count returns correctly.

The standard deviation and max do not seem sensible. Let's investigate the column.

# find products with high prices
prods[prods['prices'] > 30]

These are clearly erroneous.
We will replace them with the mean price of their department (without those values).

# remove prices for these products
prods.loc[prods['product_id'].isin(['21553', '33664']), 'prices'] = None

prods[prods['product_id'].isin(['21553','33664'])]

# take subset of department 16
dept16 = prods[prods['department_id'] == '16' ]
dept16

# find mean of prices for department 16
dept16['prices'].mean()

# impute this value for the erroneous (now null) prices
prods.loc[prods['product_id'].isin(['21553', '33664']), 'prices'] = 8.016652161299682

# check imputation
prods[prods['product_id'].isin(['21553','33664'])]

# find summary stats again
prods.describe()

products dataset is now clean.

## 3. Departments

# check dimensions
dept.shape

# view dataset
dept.head()

Clearly the dataframe needs to be transposed.

### 3.1 Wrangling

### 3.1.1 Transpose dataframe


# transpose dept
dept = dept.T

# reset index column
dept.reset_index()

# extract column names
header = dept.iloc[0]

# take only the dataset with the headers
dept = dept[1:]

# add headers
dept.columns = header

dept

### 3.1.2 Datatypes

# check datatypes
dept.dtypes

### 3.1.3 Columns to drop

No columns need to be dropped.

### 3.1.4 Columns to rename

No columns need to be renamed.

## 3.2 Consistency checks

### 3.2.1 Duplicates

# check for duplicate values
dept.duplicated().sum()

There are no duplicates to clean.

### 3.2.2 Missing values

# check for nulls
dept.isnull().sum()

There are no nulls to clean.

### 3.2.3 Inaccuracies

# view dataframe
dept

A scan of the department names suggests that some of the entries should be renamed:

+ 'dry goods pasta' to 'dry goods'
+ 'dairy eggs' to 'dairy'

and there is a 'missing' department.

Check with the client.

It is verified that the suggested name changes can be made. Still waiting on verification for the 'missing' department.

# Rename 'dry goods pasta' to 'dry goods' and 'dairy eggs' to 'dairy'
dept['department'] = dept['department'].replace({'dry goods pasta': 'dry goods', 'dairy eggs': 'dairy'})

# 4. Customers

# find dimensions of dataset
cust.shape

# view dataset
cust.head()

## 4.1 Wrangling

### 4.1.1 Datatypes

# check datatypes
cust.dtypes

user_id can be converted to strings, since they represent names rather than quantifiable data.

# convert user_id to string
cust['user_id'] = cust['user_id'].astype(str)

There are no further mixed datatypes to clean.

### 4.1.2 Columns to drop

First name and surname columns are PII and should be considered for removal from the analysis.

Verification with the client confirms that these columns can be removed from the analysis.


# drop first name and surname columns
cust = cust.drop('Surnam', axis=1)
cust = cust.drop('First Name', axis=1)

### 4.1.3 Columns to rename

# list the columns
cust.columns

Rename columns to all lower case to be consistent with formatting across the script. Correct misspelling of 'n_dependants' to 'n_dependents'. Rename 'fam_statust' to 'family_status' to be more intuitive.

# rename columns as described above
cust.rename(columns={'Gender': 'sex'}, inplace=True)
cust.rename(columns={'STATE': 'state'}, inplace=True)
cust.rename(columns={'Age': 'age'}, inplace=True)
cust.rename(columns={'n_dependants': 'n_dependents'}, inplace=True)
cust.rename(columns={'fam_status': 'family_status'}, inplace=True)

## 4.2 Consistency checks

### 4.2.1 Duplicates

# check for duplicate values
cust.duplicated().sum()

There are no duplicate rows.

### 4.2.2 Missing values

# check for nulls
cust.isnull().sum()

There are no nulls to fix.

### 4.2.3 Inaccuracies

# find summary statistics for dataset
cust.describe()

The counts for each variable are correct and there are no suspicious summary statistics. Take the dataset to be accurate.

# Export cleaned dataframes

# export dataframes
ords.to_csv(os.path.join(path,'Prepared Data','orders_cleaned.csv'))
prods.to_csv(os.path.join(path,'Prepared Data','products_cleaned.csv'))
dept.to_csv(os.path.join(path,'Prepared Data','departments_cleaned.csv'))

cust.to_csv(os.path.join(path,'Prepared Data','customers_cleaned.csv'))
