==> 03 Scripts/2. Instacart - Combining Datasets.ipynb <==
# Import libraries & dataframes

# import libraries
import pandas as pd
import numpy as np
import os

path = r'C:\Users\muril\Data Analysis Projects\07-2023 Instacart Basket Analysis\02 Data'

# Import dataframes
df_ords_prior = pd.read_csv(os.path.join(path, 'Original Data', 'order_products__prior.csv'), index_col=False)

df_prods = pd.read_csv(os.path.join(path,'Prepared Data','products_cleaned.csv'), index_col=0)

df_ords = pd.read_csv(os.path.join(path, 'Prepared Data', 'orders_cleaned.csv'), index_col=0)

cust = pd.read_csv(os.path.join(path, 'Prepared Data', 'customers_cleaned.csv'), index_col=0)

reg = pd.read_csv(os.path.join(path, 'Original Data', 'regions.csv'))

df = pd.read_csv(os.path.join(path, 'Prepared Data', 'instacart_wrangled.csv'))

# 4. Merge dataframes

## 4.1 Merge orders and orders_prior

# Check df_ords_prior shape
df_ords_prior.shape

# Check df_ords shape
df_ords.shape

# Merge dataframes
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_merged_large.head()

# Check merge
df_merged_large['_merge'].value_counts()

We cannot conclude here that every observation was merged between the two dataframes, despite the 0 counts for left and right only. Since it was an inner join, by definition the left and right counts would have to be 0. Merging by outer would show the true number of mismatching records.

# Check dimensions of merged dataframe
df_merged_large.shape

# drop _merge column
df_merged_large = df_merged_large.drop('_merge', axis=1)
df_merged_large.head()

## 4.2 Merge products

# check dimensions
df_prods.shape

df_merged_large.shape

# merge dataframes
df_merged = df_merged_large.merge(df_prods, on='product_id', how='left'

df_merged.head()

df_merged.isnull().sum()

There are nulls in the column that we expect there to be nulls, as in the previous scripts.

df_merged.shape

## 4.3 Merge customers

# list columns of dataframes
df_merged.columns

cust.columns

del df_ords
del df_prods
del df_ords_prior

# merge customers with df_merged
df_merged = df_merged.merge(cust, on='user_id', how='left')

# check for nulls 
df_merged.isnull().sum()

## 4.4 Merge regions

# merge regions
df.merge(regions['state','region'], on='state', how='left')

# Export dataframe

# Export data to csv
df_merged.to_csv(os.path.join(path, 'Prepared Data', 'instacart_wrangled.csv'))

# Export data to pkl
df_merged.to_pickle(os.path.join(path,'Prepared Data', 'instacart_wrangled.pkl'))
