==> 03 Scripts/3. Instacart - Labels & Flags.ipynb <==
# Import libraries & dataframes

import pandas as pd
import numpy as np
import os

path = r'C:\Users\muril\Data Analysis Projects\07-2023 Instacart Basket Analysis\02 Data'
ords_prods_merged = pd.read_csv(os.path.join(path, 'Prepared Data', 'orders_products_merged.csv'), index_col=0)

cust = pd.read_csv(os.path.join(path, 'Prepared Data', 'customers_cleaned.csv'), index_col=0)

# 5. Assign labels

## 5.1 price_label flag

# assign high, mid-range and low flags to products
ords_prods_merged.loc[ords_prods_merged['prices'] > 15, 'price_range_loc'] = 'High-range product'
ords_prods_merged.loc[(ords_prods_merged['prices'] <= 15) & (ords_prods_merged['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 
ords_prods_merged.loc[ords_prods_merged['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

# Check count of loc function output
ords_prods_merged['price_range_loc'].value_counts(dropna=False)

## 5.2 Busiest days flag

# find the counts of orders by day of the week
ords_prods_merged['order_day_of_week'].value_counts(dropna = False)

Saturday (0) and Sunday (1) are the busiest days and Wednesday (4) and Tuesday (3) are the quietest days.

# add column to dataframe that tags the busiest day and the quietest day

result = []

for value in ords_prods_merged["order_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

# add this column to the dataframe
ords_prods_merged['busiest_day'] = result

ords_prods_merged['busiest_day'].value_counts(dropna = False)

# add column to dataframe that tags the busiest two days and the quietest two days

result2 = []

for value in ords_prods_merged["order_day_of_week"]:
  if value in (0,1):
    result2.append("Busiest days")
  elif value in (3,4):
    result2.append("Slowest days")
  else:
    result2.append("Regularly busy")

# add this column to the dataframe
ords_prods_merged['busiest_days'] = result2

ords_prods_merged['busiest_days'].value_counts(dropna = False)

The function seems to have performed correctly, the totals are correct and the entries return correctly for the observations above.

## 5.3 Busiest time of day flag

# frequencies of order hour
ords_prods_merged['order_hour_of_day'].value_counts(dropna = False)

The time periods “Most orders”, “Average orders” and “Fewest orders” will be distributed by 25%, 50% and 25% respectively of the number hours of the day, meaning that Most Orders will be the top 6 hours, Fewest Orders will be the bottom 6 hours and Average Orders will be the middle 12 hours.

Most Orders hours - 10, 11, 14, 15, 13, 12

Fewest Orders hours - 0, 1, 5, 2, 4, 3

# add column to dataframe that tags the busiest and quietest hours

result3 = []

for value in ords_prods_merged["order_hour_of_day"]:
  if value in (10, 11, 14, 15, 13, 12):
    result3.append("Most orders")
  elif value in (0, 1, 5, 2, 4, 3):
    result3.append("Fewest orders")
  else:
    result3.append("Average orders")

# add this column to the dataframe
ords_prods_merged['busiest_period_of_day'] = result3

# check flag counts
ords_prods_merged['busiest_period_of_day'].value_counts(dropna = False)

del result
del result2
del result3

## 5.4 Loyalty flag

# find maximum number of orders per user
loyalty = ords_prods_merged.groupby('user_id')['order_id'].nunique().reset_index()
loyalty.head()

# assign loyalty flag - greater than 40 returns 'loyal', less than 10 returns 'new', otherwise 'regular'
loyalty['loyalty_flag'] = loyalty['order_id'].apply(lambda order_id: 'loyal' if order_id > 40 else 'new' if order_id < 10 else 'regular')

loyalty['loyalty_flag'].value_counts()

loyalty['loyalty_flag'].isnull().sum()

Flag assignation appears to have functioned correctly.

# merge with dataframe
ords_prods_merged = ords_prods_merged.merge(loyalty[['user_id','loyalty_flag']], on='user_id', how='left')

del loyalty

ords_prods_merged.head()

# Export dataframe
